Crate solana_program_pack

The Pack serialization trait This is a specific serialization API that is used by many older programs in the Solana Program Library to manage account state. It is not generally recommended for new code since it does not define a language-independent serialization format.

Traits

IsInitialized

Check if a program account state is initialized

Pack

Safely and efficiently (de)serialize account state

Sealed

Implementors must have a known size